Add a subject validation in PermissionRule
Currently the interface for a Permission Rule accept conditions but do
not apply the conditions when the given subject is a class.

This issue leads to undesired behaviour where the consumer of this gem
may define a permission_rule with conditions and passing a class
expecting the conditions to work but it will return true.

This commit adds a validation that raises an error when a
permission_rule with some conditions are given a class instead of an
instance.